+++ src/ashley/machina_extensions/forum_search/forms.py
@@ -3,6 +3,7 @@
     ==================
     This module defines forms provided by the ``forum_search`` application.
 """
+from django.utils.translation import gettext_lazy as _
 from machina.apps.forum_search.forms import SearchForm as MachinaSearchForm
 from machina.core.db.models import get_model
 from machina.core.loading import get_class
@@ -13,28 +14,33 @@ PermissionHandler = get_class("forum_permission.handler", "PermissionHandler")
 
 class SearchForm(MachinaSearchForm):
     """
-    Override Django Machina's SearchForm to allow searching poster user names
+    Allows to search forum topics and posts and allows searching poster user names
     even with an empty search in the main field.
     """
 
     def __init__(self, *args, **kwargs):
-        """
-        Init form, code based on base class MachinaSearchForm. A filter on lti_context
-        has been added.
-        """
-        # Loads current lti_context to filter forum search
-        lti_contexts = kwargs.pop("lti_context", None)
-        super().__init__(*args, **kwargs)
         user = kwargs.pop("user", None)
+        lti_contexts = kwargs.pop("lti_context", None)
+
+        # pylint: disable=bad-super-call
+        super(MachinaSearchForm, self).__init__(*args, **kwargs)
+
+        # Update some fields
+        self.fields["q"].label = _("Search for keywords")
+        self.fields["q"].widget.attrs["placeholder"] = _("Keywords or phrase")
+        self.fields["search_poster_name"].widget.attrs["placeholder"] = _("Poster name")
+
         self.allowed_forums = PermissionHandler().get_readable_forums(
             Forum.objects.filter(archived=False, lti_contexts=lti_contexts), user
         )
-        # self.allowed_forums is used in search method of MachinaSearchForm
         if self.allowed_forums:
             self.fields["search_forums"].choices = [
                 (f.id, "{} {}".format("-" * f.margin_level, f.name))
                 for f in self.allowed_forums
             ]
+        else:
+            # The user cannot view any single forum, the 'search_forums' field can be deleted
+            del self.fields["search_forums"]
 
     def clean(self):
         """